requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.1
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
